SNOW-1853185: Implemented comments from pr
