+++ src/main/java/net/snowflake/client/core/HttpUtil.java
@@ -587,7 +587,7 @@ public class HttpUtil {
    * @param retryTimeout retry timeout
    * @param authTimeout authenticator specific timeout
    * @param socketTimeout socket timeout (in ms)
-   * @param retryCount retry count for the request
+   * @param retryCount max retry count for the request - if it is set to 0, it will be ignored and only retryTimeout will determine when to end the retries
    * @param injectSocketTimeout injecting socket timeout
    * @param canceling canceling?
    * @param ocspAndProxyKey OCSP mode and proxy settings for httpclient
@@ -630,7 +630,7 @@ public class HttpUtil {
    * @param retryTimeout retry timeout
    * @param authTimeout authenticator specific timeout
    * @param socketTimeout socket timeout (in ms)
-   * @param retryCount retry count for the request
+   * @param retryCount max retry count for the request - if it is set to 0, it will be ignored and only retryTimeout will determine when to end the retries
    * @param ocspAndProxyAndGzipKey OCSP mode and proxy settings for httpclient
    * @return response
    * @throws SnowflakeSQLException if Snowflake error occurs
@@ -661,13 +661,14 @@ public class HttpUtil {
    * @param retryTimeout retry timeout
    * @param authTimeout authenticator specific timeout
    * @param socketTimeout socket timeout (in ms)
-   * @param retryCount retry count for the request
+   * @param retryCount max retry count for the request - if it is set to 0, it will be ignored and only retryTimeout will determine when to end the retries
    * @param ocspAndProxyAndGzipKey OCSP mode and proxy settings for httpclient
    * @param retryContextManager RetryContext used to customize retry handling functionality
    * @return response
    * @throws SnowflakeSQLException if Snowflake error occurs
    * @throws IOException raises if a general IO error occurs
    */
+  @SnowflakeJdbcInternalApi
   public static String executeGeneralRequest(
       HttpRequestBase httpRequest,
       int retryTimeout,
@@ -700,7 +701,7 @@ public class HttpUtil {
    * @param retryTimeout retry timeout
    * @param authTimeout authenticator specific timeout
    * @param socketTimeout socket timeout (in ms)
-   * @param retryCount retry count for the request
+   * @param retryCount max retry count for the request - if it is set to 0, it will be ignored and only retryTimeout will determine when to end the retries
    * @param httpClient client object used to communicate with other machine
    * @return response
    * @throws SnowflakeSQLException if Snowflake error occurs
