+++ src/test/java/net/snowflake/client/core/SessionUtilLatestIT.java
@@ -643,37 +643,4 @@ public class SessionUtilLatestIT extends BaseJDBCTest {
       assertEquals((int) ErrorCode.IDP_INCORRECT_DESTINATION.getMessageCode(), ex.getErrorCode());
     }
   }
-
-  @Test
-  public void testOktaAuthRequestsAreRetriedUsingLoginRetryStrategy() {
-    final String oktaSSOAuthPath = "api/v1/authn";
-    final String oktaTokenAuthPath = "app/snowflake/tokenlikepartofurl/sso/saml";
-    List<String> oktaAuthURLs = new ArrayList<String>();
-    oktaAuthURLs.add("https://anytestpath.okta.com"); // default *.okta.com URL
-    oktaAuthURLs.add("https://vanity-url.somecompany.com"); // some custom Vanity OKTA URL
-
-    for (String oktaAuthURL : oktaAuthURLs) {
-      try {
-        // Check that SSO path is recognized as the new retry strategy
-        assertThatPathIsRecognizedAsNewRetryStrategy(oktaAuthURL, oktaSSOAuthPath);
-        // Check that Token path is recognized as the new retry strategy
-        assertThatPathIsRecognizedAsNewRetryStrategy(oktaAuthURL, oktaTokenAuthPath);
-      } catch (URISyntaxException e) {
-        fail(
-            "Test case data cannot be treated as a valid URL and path. Check test input data. Error: "
-                + e.getMessage());
-      }
-    }
-  }
-
-  private void assertThatPathIsRecognizedAsNewRetryStrategy(String uriToTest, String pathToTest)
-      throws URISyntaxException {
-    URIBuilder uriBuilder = new URIBuilder(uriToTest);
-    uriBuilder.setPath(pathToTest);
-    URI uri = uriBuilder.build();
-    HttpPost postRequest = new HttpPost(uri);
-    assertThat(
-        "New retry strategy (designed to serve login-like requests) should be used for okta authn endpoint authentication.",
-        SessionUtil.isNewRetryStrategyRequest(postRequest));
-  }
 }
